Migrate from swarm to compose since we can now count on a cloud environment.
Docker swarm is great if you need multiple machines to coordinate but want something lighter weight than kubernetes.  Docker compose is what you're looking for if you want something to run only on one machine but want to be able to inspect it as it goes.

Before, we thought we'd need to run on-prem, but we are newly confident we can count on a cloud provider, so it's better to use docker compose instead since it's much simpler.