Around 10 am today, our docker builds started failing. (#1212)
we have no root cause for this, but it seemed to be related to finding application.conf.
a search of the internet suggests that play's many mechanisms for building a
distributable version of an application often fail to find application.conf.
We don't know why it failed in this case, but the play docs recommend providing
the config.file parameter on the command line just in case.  It happens to work.