Stop storing blank values in applicant data. (#1451)
* Stop storing blank values in applicant data.

If updates are empty, they
are not put into the applicant data json, but metadata is still
updated. Arrays (enumerator and multi select questions) with empty
updates get deleted, but the metadata is still updated.

Questions are "answered" if any of their scalars have a value in
applicant data. There might be weird behavior with old applicant data
that is storing blank values, but blank values shouldn't exist in the
future.

Errors for multi-field questions (address and name) for each of their
fields are present if the question is answered, not just if the field is
answered.

* Required file questions cannot be skipped

* ApplicantData#maybeClearRepeatedEntities returns a bool

* Add required validation

* Delete enumerator metadata if it is the only thing stored there before an update

* Empty update means delete stored scalars.

* Delete enumerator if it empty before updates. The other way would delete anything stored there.

* add tests

* Use Optional<Tag> over null for views.

* add a comment

* Can't use condWith with optionals because it doesn't lazily evaluate them.

* Update universal-application-tool-0.0.1/app/services/applicant/ApplicantServiceImpl.java

Co-authored-by: Caroline Danzi <danzicr@miamioh.edu>

* Update universal-application-tool-0.0.1/app/services/applicant/question/ApplicantQuestion.java

Co-authored-by: Caroline Danzi <danzicr@miamioh.edu>

* carolines suggestions

Co-authored-by: Caroline Danzi <danzicr@miamioh.edu>