Trigger Azure file upload off of form submit rather than clicking the "Upload" button.
The previous behavior caused file upload validation from validation.ts to not trigger. Since that validation didn't occur, empty uploads would cause an exception to be thrown when attempting to access the file type. Triggering off of form submit allows the validation to occur first.

This necessitated updating the `/dev/fileUpload` handler for Azure to put the form inside of a div.