Remove shouldDisplayQuestionErrors / ApplicantQuestion.getQuestionErrors.
The former wasn't necessary since every time this was set to false, it was due to wanting to show a field-specific error message. As part of this, I also ensured that field-specific errors were always returned in getAllTypeSpecificErrors (test to be updated next).

Also removed ApplicantQuestion.getQuestionErrors, since it only had a single callsite and was purely passhtrough to the question renderer.