ApplicationQuestionRenderer refactoring to make subclasses responsible for providing a method to render form content (#2318)
* Inverting ApplicantQuestionRenderer pattern from having implementations call renderInternal, which creates the wrapping div to instead have them implement renderFTag.

Why? Simpler to ensure all input-providing question types are wrapped in the same div.

* Remove shouldDisplayQuestionErrors / ApplicantQuestion.getQuestionErrors.

The former wasn't necessary since every time this was set to false, it was due to wanting to show a field-specific error message. As part of this, I also ensured that field-specific errors were always returned in getAllTypeSpecificErrors (test to be updated next).

Also removed ApplicantQuestion.getQuestionErrors, since it only had a single callsite and was purely passhtrough to the question renderer.

* Fixing test assertions.

* bin/fmt

* Making constructor and class package private.