Required questions now return validation errors if left blank.
Why? Consolidates validation of a given question.

If only QuestionImpl were changed, errors would be presented the first time a required question was rendered, which is undesirable. Instead, propagate whether a question is being rendered as part of initial display to the user or in response to a failed attempt to POST the block. To ensure consistent rendering behavior, the responsibility of calling Question.getValidationErrors was moved to the base Renderer class.