Redirect to draft question when trying to edit published. (#2915)
It's possible for global admins to edit question in parallel: in one tab create a new version of an existing question and then, in the second tab, try to edit the original question. Currently if that happens the admin sees the original question data even though a new draft version exists. This PR handles this case and redirects admin to the draft version.