Remove auto-approve from setup.py (#3028)
Currently setup.py auto-approves and applies any changes when mode is staging or prod. It's dangerous as user should review those changes before applying them. The auto-approve seems to have been added for automation (CI tests) which currently are not implemented. In future, when we implement the tests, we can add this option back guarded by some flag or environment variable used only in tests.