Move ADFS_DISCOVERY_URI and AWS_REGION to civiform_config.sh (#3048)
ADFS_DISCOVERY_URI is currently a secret. It should be a regular variable instead, similar to APPLICANT_OIDC_DISCOVERY_URI.

AWS_REGION is currently hardcoded to us-east-1 and user can't change it without updating variables.tf which should be considered immutable to deployers. This PR reads AWS_REGION from the civiform_config.sh if it's present and defaults to us-east-1 if it's absent.