Ask user for secret values during initial AWS deployments (#3093)
After the initial AWS deployment we create secrets with empty values in AWS secret manager. CiviForm app requires non-empty values and won't start with empty. This PR adds a code that asks user to provide values for the secrets and sets them in the secret manager. The code doesn't ask for values if they secret is already non-empty.