Change DB password after initial terraform setup is done. (#3120)
Original DB password generated by the Terraform is stored in the Terraform state. We want to avoid exposing such info in the state even though access to the state should be limited. To achieve that we now regenerate password. Update database to use the new password, update the secret value and restart ECS service to pickup the new secret value.