Enable fingerpriting and caching for static files even on dev mode (#3394)
This improves speed of browser tests. For quesiton_lifecycle.test.ts duration goes from 250s to 140s. Fingerprinting guarantees that when while changes - a new version will be requested. The downside of that approach is that it adds an extra step (fingerprinting) into incremental rebuild of static assets (TS files). But that adds only around 1-2s so it's practically unnoticeable. It might become an issue in future if we have bigger static files, but with current guidelines (minimum JS) that unlikely to happen soon.