Render "referenced programs" text outside of active/draft rows (#3405)
Relation between question and programs that reference it is orthogonal to the question being active or draft. Programs include question by name and it's not possible to include, for example, only draft version or only active version of a question. This PR makes it more clear by moving "referenced programs" text outside of active/draft rows.