Update program creation / editing forms to display internal program name as a link (#3441)
### Description

This reorders the fields in the program creation / editing forms to emphasize the external-facing text. It further deephamsizes the internal-facing name to instead be the "program URL" (not editable after program creation) and the internal-facing description to be the admin note.

This is stage 2 of the changes outlined in https://github.com/civiform/civiform/issues/3085#issuecomment-1246008757.

Lastly, I'll enforce that the provided input for new programs be of the "slug" form. That'll involve a fair amount of browser test changes so I'll do that separately.

### Screenshots

New program creation:
<img width="1549" alt="Screen Shot 2022-09-19 at 11 03 11 AM" src="https://user-images.githubusercontent.com/1870301/191084075-9fdb4f2c-648a-49e4-a5f1-dae48174a6ce.png">

Editing an existing program (program URL editing is disabled):
<img width="1141" alt="Screen Shot 2022-09-16 at 12 20 53 PM" src="https://user-images.githubusercontent.com/1870301/190715047-d4f71a79-65a5-4d56-adf9-d5dcfede06f3.png">

## Release notes:

Reorder fields for CiviForm admins creating / editing programs to better reflect how they'll be used.

### Checklist

- [X] Added the correct label: < feature | bug | dependencies | infrastructure | ignore-for-release >
- [X] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

### Issue(s)

#3085