Showing user-facing text in the visibility condition builder (#3500)
### Description

In addition to showing the user-facing text in the visibility condition builder, we'll also to show the internal-facing identifier within the "Admin ID:" field (similar to #3498) in order to allow the CiviForm admin to differentiate questions with the same user-facing text.

Additionally, we dynamically build a string representation of the visibility condition. Here, I thought the internal facing identifier made more sense to use since it's not necessarily a multi-sentence value.

### Screenshots

Note: Added screenshot tests for the show/hide visibility conditions. Leaving snapshots of the before/after here for posterity.

Before:
![Screen Shot 2022-09-26 at 10 14 59 AM](https://user-images.githubusercontent.com/1870301/192339584-577b68e0-704b-4f55-a4a4-11e58b4bd5e1.png)

After:
![Screen Shot 2022-09-26 at 10 14 27 AM](https://user-images.githubusercontent.com/1870301/192339514-1a8b1190-38cc-4e3e-8ea1-37d0480bbcae.png)

## Release notes:

Surfacing the user-facing text within the CiviForm Admin's visibility condition builder UI.

### Checklist

- [X] Added the correct label: < feature | bug | dependencies | infrastructure | ignore-for-release >
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

### Issue(s)
#3075