Fix login / select language page titles to be more accessible (#3549)
### Description

#### Login page
We had logic to suffix the page title with text depending on the layout (e.g. `- CiviForm` for applicants and `- CiviForm Admin Console` for program/CiviForm admins).

Our login page wasn't make use of either of those specific layouts. This pushes the "suffix the page title" logic down to the `BaseHtmlLayout` class with a default value of "CiviForm".

Note: While here, I also filed #3548 to track the fact that the "Login" text is not localized.

#### Select language
This had an incorrect set of text. Since it wasn't localized, I also removed it from the translations files and inlined it directly with a comment. Finally, I moved the `<h1>` SR-only tag above the content that it was describing for clarity.

## Release notes:

Updating the titles for the login page and language selection page to be more accessible.

### Checklist

- [X] Added the correct label: < feature | bug | dependencies | infrastructure | ignore-for-release >
- [X] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

### Issue(s)

Fixes #3545