Fail on new images being added (#3551)
jest-image-snapshot automatically adds image snapshot if it's missing during test run. That's good for local development but for CI it's undesired: we want test to fail in cases of missing snapshots. This PR adds extra step to github action workflow which ensures that no screenshots being added after test run.