Making cloned enumerator field entries have unique IDs (#3585)
### Description

We generate a "template" for the enumerated entity server-side. Previously this was generated a random ID. However, when the JavaScript in main.ts cloned the node, it also carried the ID along, generating a non-unique ID.

This keeps a global counter and replaces the ID each time a new entity is added.

Also updated integration tests to test these scenarios as well. I manually validated that it failed when I:
  * Didn't update the input element's ID
  * Didn't update the `label` element's `for` attribute
  * Didn't update the error div's ID

## Release notes:

Fix accessibility violations when applicants were adding answers to an enumerated question.

### Checklist

- [X] Added the correct label: < feature | bug | dependencies | infrastructure | ignore-for-release >
- [X] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

### Issue(s)

Fixes #3565