Heading A11Y fixes for preview, review, and block editing pages (#3579)
### Description

While here, I discovered that we don't localize the "1 of 5" text shown in the applicant block editing form or the page titles. I filed #3578 for this.

The screenreader heading was removed for the block editing page by putting all of the "<Program name> 1 of 5" text inside of a single H1 and using spans with different classes to make sure it rendered the same as before. This way screenreaders will see all of the text as part of the h1 tag.

## Release notes:

Fixing heading tags in the applicant-facing preview, review, and screen editing pages to improve use with screen readers.

### Checklist

- [X] Added the correct label: < feature | bug | dependencies | infrastructure | ignore-for-release >
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary

### Issue(s)

Fixes #3554; Fixes #2640; #3578