Fix provider name for Auth0Provider (#3633)
I made a mistake in the original implementation. In the buggy implementation we were reading non-existing "Auth0" config setting and it fallback to "OidcClient". In the new we always return "Auth0". This change is safe as Auth0 is used only by our staging and I updated Auth0 application config to include new redirect URI.