Extract block question validation logic into separate utility class (#3736)
ProgramBlockValidation contains logic that checks whether a question can be added to a block. The class is introduces mainly for the purpose of sharing between two callsites:

During question bank rendering to ensure that only eligible questions are rendered for a given block.
When processing request for adding question to a block to guard against bad request or bugs in our "client-side" checks at step 1.
Additionally PR replaces DuplicateQuestionException with a more generic CantAddQuestionToBlockException that can provide few different reasons why question cant be added.