Fix scrolling issue with question bank question (#3754)
When creating a new question from the question bank, the bottom-most question is not visible. This is due implementation details of question bank where bottom 48px are always below the screen. To fix it we add 48px padding to the question type list to make all questions visible.

I didn't add/update any tests. I tried using playwright's isVisible() to check whether the bottom-most question is visible but it returns true even if it's below the screen.