Remove noisy " Application {} does not include an applicant name" warning (#3756)
During browser test run there is a lot of warnings in console like

[warn] s.a.ApplicantData - Application {} does not include an applicant name. This is expected for guest users.
They are misleading as they don't indicate that something is wrong or should be fixed. Additionally it uses incorrect template with missing application id. Reading the code there is no application id attached to ApplicantData as it is expected to contain answers to all questions. That's why I decided to remove this warning message altogether as it doesn't give any useful information.