Fix program card external link rendering (#3815)
It got broken by #3805 The reason is that link produced by LinkElement now always use display: inline-flex so that element inside the link (text and icon) are aligned properly. But it means that link is always inline. If callers want to make it block - they need to wrap it with div() as callers can't override LinkElement class.