Add currency to predicate options (#3858)
Allows currency question types to be used in Predicates.  Currently they are selectable but have no operators.

Note: currency is stored as a cents value. This PR takes the user provided predicate dollars/cents value and converts it to cents for storage, and likewise converts cents to dollars/cents when displayed.

`PredicateValue.toDisplayString()` is updated to now require the question the predicate is on.  The method only has one caller and it now directly handles the case where the question is not available.  (This situation is likely indicative of a coding/data bug though)