Explicitly wait for browser to end up on login page after logout (#3865)
Until now we've been relying on implicit navigation event wait. That wait is added by page.click() and assumes that the next page that will be fully loaded is the loginForm page. That assumption comes from the fact that logout usually involves a series of redirects that end up on loginForm and that all redirects are server-side using 3xx responses. In Seattle staging one of those redirects is implemented client-side and that page fires navigation event violating our expectation.

In this PR I explicitly wait for browser to end up on the /loginForm page.