Refactor JS files to use consistent initialization pattern (#3869)
Most JS files today use initialization pattern where they check whether certain DOM element is present on the page and if so - initialize themselves. This PR updates all files to use the same pattern. This is required for upcoming JS bundling change where we'll compile all files together for admins and for applicants. So each JS file must handle the case where it's loaded on a page that doesn't actually "need" that file.