Rename only use of ViewOnly to common ReadOnly (#3894)
The code base commonly uses ReadOnly to indicate uneditable things. This renames the only use of ViewOnly to that.

All of our UI rendering classes are called View regardless of if the content is editable or not so using ViewOnly for a different use of "View" in them is confusing.