Add flag to disable browser error watcher (#3907)
It is useful for staging prober tests where we don't control some sites such as auth providers and they might cause JS or network errors that don't fail test but trigger the error watcher. The main goal of brower error watcher to catch errors in civiform and that is covered well enough by browser tests.