Add `use_js_bundles` feature flag (#3909)
When the flag is enabled JS bundles will be served instead of individual JS files. The plan is to add the flag, flip the default value to true, ensuring that all tests pass, and wait for a release or two. After that I'll clean up the flag and the old approach.