Add eligibility feature (#3932)
Adds the ability to specify predicates on blocks that are used to determine if the applicant is eligible for the program.

If an applicant is not eligible they are shown a screen indicating that and can not proceed.

Setting/removing/evaluating eligibility is very similar to the existing visibility feature. The Admin management of them is side-by-side in the block UI and is mostly the exact same View code.

Eligibility of an applicant is done upon submit of the block. Evaluation of eligibility is conceptually done "after" the block and before determination of the next "visible" block to show.

The feature is flag guarded for both seeing the admin UI and evaluation of any set eligibility predicates.