Enable typescript strict checks on server code (#3999)
The main change is that accessing (calling methods or accessing properties) on nullable variables is now an error. Developer either have to explicitly check that value is not-null or add non-null assertion. For that I introduced assert() helper function that ensures that provided value is non-null and throws error otherwise. It required me to add it in a lot of places, mainly around querySelector() and getElementById() calls. It's quite annoying but provides extra safety.

We can also implement something like safeQuerySelector described here: https://effectivetypescript.com/2020/07/27/safe-queryselector/ But I decided to go simpler, more universal approach (we'd need to have assert() for other cases as well).