Add validation check upon application submit (#4012)
Upon submit verify that all questions in the application have been answered. If any have not, show the review page with an appropriate toast message.

This situation will occur if:

The application has been filled out and the user is on the page with the Submit button
In another tab they change the answer to a visibility question in a way that should show more screens in 1
Return to the tab and click submit.