Fix vscode-setup
Previously, bin/sbt was not taking any additional arguments and passing them along to docker::dev_and_test_server_sbt_command. This meant that when vscode-setup tried running 'bin/sbt makePom', it wouldn't actually do that and it would just sit in the shell. Also, for some reason, docker compose can't seem to figure out what the 'civiform-shell' service is, even though it's defined in docker-compose.dev.yml.  However, we're just using the civiform service anyway, so this fixes the cp command to use the civiform service to copy the target directory to the local server directory so we can munge the pom file appropriately.