Fixes for bin/run-browser-tests-local
This fixes several issues found with running PWDEBUG=1
bin/run-browser-tests-local.

1. This adds a --local flag to bin/run-browser-test-env, which sets
   BASE_URL to http://localhost:9999, which the local browser tests
expect.  Otherwise, upon logout, it will attempt to go to civiform:9000
instead and fail.

2. This exposes port 3390 on the dev-oidc container. This is necessary
   for the app to be able to redirect there when running the tests
locally. Note that this will likely collide with a running instance of
bin/run-dev, and we can probably fix things up somehow to have it point
to a different port, but this works for a first pass.

3. This adds the flag for starting the JVM debugger on port 9457 to
   allow connecting the debugger during browser tests.

4. This updates the test_oidc_provider.js file to add the paths for
   localhost:9999 as allowed redirect URIs. The container will need to
be rebuilt and pushed once this is merged.