Allow for no address correction suggestions
To handle the case where the ESRI server doesn't return any address correction suggestions, or an error occurs, this changes EsriClient.getAddressSuggestions to always return an AddressSuggestionGroup. If we couldn't get any suggestions from the server, it will create an AddressSuggestionGroup containing the original address, so that we can render this in the view. This also changes the view to handle when we get no suggestions. It will show the address entered and ask the user to edit it to correct it, or allow the user to save the inputted address anyway and continue.