Temporary fix for program index page
Because some of the programs are not in the active version, and we need to sync the questions for each program to calculate eligibility state, we must sync each program with a version it is associated with. This is a quick fix to utilize the ReadOnlyVersionedQuestionService instead of the ReadOnlyCurrentQuestionService.