Add "Learn more" link to address correction tooltip
This adds a "Learn more" link to the address correction tooltip when AC
is not enabled, pointing to the docs page that describes how to enable
it. It also reduces the top margin to zero in this case so that the
tooltip doesn't disappear when a user mouses away from the tooltip icon
to go to click the link.