Refactor email notification functions and change TI email behavior
This refactors the 'maybeNotify' functions that email the applicants and TIs. This combines each into one function, rather than separate functions depending on if a default status exists. Additionally, this sends the status email body to the TI as well as the applicant, rather than the default 'application received' email to the TI.