Remove async version of getReadOnlyQuestionService
As far as I can tell, there isn't a good reason to have an async version of this function. It was just providing a CompletableFuture.completedStage wrapper around the synchronous version of this method, which is really just doing it synchronously. It was being used to chain other async blocks, but we can use CompletableFuture.supplyAsync instead.