Fix admin_settings test for probers
Because we have ESRI_ADDRESS_CORRECTION_ENABLED on staging, while this test would pass in the hermetic test env, it would fail on staging since it would be trying to enable something that's already enabled. This switches it to use the CF_OPTIONAL_QUESTIONS setting instead, since this should probably always be enabled, and to test disabling, re-enabling, then saving with it still enabled.