Fix use of quotes in predicates
When saving a string used in a predicate, we must wrap double quotes around the string in order for it to be handled correctly. When editing an existing predicate, it would render the string with the double quotes intact. This meant that when it saved the predicate, it would wrap it again in double quotes. This would cause an unhandled exception when a user went to apply for a program.

With these changes, we now render the predicate string without double quotes. Then upon saving, we strip the value of all double quotes, then wrap it with double quotes. This should ensure the value is always in the format of "value" and won't result in JSON pathing errors.