Only set submit time on previous applications that don't have it
In order to address https://github.com/civiform/civiform/issues/3227, we
were iterating through all previous applications that were not marked
OBSOLETE and setting the submit time. However, we had not obsoleted the
previous, current ACTIVE application at this point, so the previous
revision of the application would always get its submission time
updated, causing problems with metrics. This changes the logic to only
set the submit time on a currently ACTIVE application if it doesn't
already have a time set.