Fix removing UNIVERSAL tag from new draft question
When creating a new draft of an already-published question where the question currently has the UNIVERSAL tag and now we're trying to remove it, the tag would not actually get removed because we copy over all tags from the old version of the question to the new one. Since the UNIVERSAL tag is managed by the state of the universal field in the QuestionDefinition, this removes the tag if the new definition does not have universal set.