Rename StorageClient to ApplicantStorageClient (#6041)
### Description

Right now, the only files that CiviForm stores are files uploaded by applicants to answer questions. We use `StorageClient` interface and its cloud-storage-provider-specific implementations to do that file upload.

For the new program images feature, we'll now also allow admins to upload program image files and store them in a publicly accessible location. This PR renames the current applicant file upload classes to have `Applicant` in the name so that it's clear in the code that these classes are for applicant files. A future PR will add a `PublicStorageClient` interface for the program images.

This PR also renames the implementations of `StorageClient` to have the cloud provider's name in them:
 - `BlobStorage` -> `AzureApplicantStorage`
 - `SimpleStorage` -> `AwsApplicantStorage`

This makes it easier to find all the Azure-specific and AWS-specific storage code.

## Release notes

### Checklist

#### General

Read the full guidelines for PRs [here](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#creating-a-pull-request)

- [x] Added the correct label: < feature | enhancement | bug | dependencies | infrastructure | ignore-for-release | database >
- [x] Assigned to a specific person, `civiform/developers`, or a [more specific round-robin list](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-reviewers).
- [x] Performed manual testing (Chrome and Firefox if it includes front-end changes)

### Instructions for manual testing

 Run `bin/run-dev`  and verify you can upload a file as an applicant.

Note: Our dev Azure setup doesn't currently work, so we can't use `bin/run-dev –-azure` to verify applicant file upload with Azure.

### Issue(s) this completes

#5676