[Program Card Images] Add a "Delete image" option for admins. (#6314)
### Description

This PR allows admin to delete the current program image for any given program. A "Delete image" button is added to the program image edit page. When that button is clicked, a confirmation modal appears. If the user confirms in that modal, the program definition is updated to no longer have an image file key stored.

In this PR, the image filekey is removed from the database, but the image itself is *not* removed from cloud storage. The next PR *will* remove the image from cloud storage.

## Release notes

### Checklist

#### General

Read the full guidelines for PRs [here](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#creating-a-pull-request)

- [x] Added the correct label: < feature | enhancement | bug | dependencies | infrastructure | ignore-for-release | database >
- [x] Assigned to a specific person, `civiform/developers`, or a [more specific round-robin list](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-reviewers).
- [x] Created unit and/or browser tests which fail without the change (if possible)
- [x] Performed manual testing (Chrome and Firefox if it includes front-end changes)

#### User visible changes

- [x] Manually tested at 200% size
- [x] Manually evaluated tab order

### Instructions for manual testing

1. Enable the PROGRAM_CARD_IMAGES flag.
2. Edit a program then click "Edit program image" on the program's edit page.
3. Verify that the "Delete image" button is initially disabled because no image was set on the program.
4. Upload and save an image.
5. When the page reloads, verify that the "Delete image" button is enabled.
6. Click "Delete image" and then click "Delete image" on the confirmation modal.
7. Verify that when the page reloads, the program card preview no longer shows an image and a success toast saying "Image removed" is shown.
8. Publish the program.
9. Logout and verify that the homepage that the applicant sees does not show the image.

### Issue(s) this completes

Part of epic #5676
Part of sub-task #6293