Plumbing for PrimaryApplicantInfoTag
This creates the PrimaryApplicantInfoTag, which is 1:1 with a
QuestionTag and includes some additional information about the tag.
This also adds a field to the QuestionDefinition to store the one or
more PrimaryApplicantInfoTags applied to the question, and adds
functions to the model and other classes to support adding and removing
the underlying QuestionTag appropriately.