Fix evolution for primary applicant info
This evolution neglected to have the IF NOT EXISTS guard for creating the extensions and columns. This can be problematic for the extension if the database user does not have permission to create extensions. In this case, the extension will need to be added manually, but then this evolution would fail because it didn't have the IF NOT EXISTS on it.