Move toggle button ID to top level component
Because this was not on the top level component of the toggle, it was possible to click outside the bounds of the toggle + label, have the toggle...toggle, but not have listeners attached to the toggle fire. Putting the ID on the top level component allows the client-side JS to catch the event wherever the user clicks.