Modify ApplicantModel and ApplicantData for Primary Applicant Info
This modifies ApplicantModel to include the new columns for Primary Applicant Info. Note that empty (`""`) or blank (whitespace only) strings will get coerced into a null value so that the getters can return the correct Optional value.  This also modifies ApplicantData and places that use it to use these columns instead of the WellKnownPaths.  The only exception is for the job that is fixing an error in the applicant data blob around date of birth, which will be removed as soon as all CEs are on a version that runs this job.

The ApplicantData object now also includes the ApplicantModel object. This was debated in the TDD if this was a good idea, but we landed on doing this to simplify code that currently only has access to the ApplicantData object, so we don't have to pass down the ApplicantModel object separately.