Refactor ProgramPredicateConfigureView.renderValueRow (#7282)
The ProgramPredicateConfigureView.renderValueRow method had forked logic based on maybeAndNode.isPresent. Instead of duplicating the logic, we can just create an empty map if maybeAndNode is not present, and handle both cases inline, so the logic is only happening once and it's easier to see how the two cases differ.