Fix add button enable/disable logic when answering enumerator questions (#7321)
When filling out an enumerator question in an application, the Add button is supposed to be disabled if any of the entries are blank. This only worked for entries that were created in the current page load. If entries were saved and then the page was reloaded, they would no longer participate in the enable/disable logic. This fixes that inconsistency so if an existing entry is blanked out the add button will be disabled.