Fix poor optional handling in ProgramPredicateConfigureView
By creating a bunch of Optionals (maybePredicateValue, maybeScalar, maybeOperator) based on the existence of the root Optional (maybeLeafNode), it obfuscates the reason that they are optional, and then required an unchecked .get() and handling of an impossible Optional.empty case.

Replace maybePredicateValue.map with maybeLeafOperationNode.map, which is the root optional, and then we know the rest of the fields exist if it does. Then inline the hasOperatorRightHandType method which now has a trivial implementation.