Initial updates for play ebean v7
Updating deprecated code

Bug fix api retired key.

It wasn't reloading the object for one.

Second there is a difference in the instant precision in the assertion when hasmessage serializes the object. Just setting the id and not the whole object to be in the exception message string. We don't need to be logging the details anyway.

Fixing metrics controller tests