Combine redundant addPredicate and addPredicates test helper methods (#7459)
The addPredicate and addPredicates test support methods in admin_predicates.ts were redundant with each other. Combine them, and use a rest parameter (...) to make the caller syntax simple for either case. This also has the benefit of labeling all the params on all callers, which was previously not readable for addPredicate.

Also waitForPageJsLoad seems to have been unnecessary since addPredicate had it but addPredicates does not.