Refactor configurePredicate in admin_predicates.ts (#7539)
The increment of groupNum should happen exactly once per loop iteration, but is executed in various places, which is flimsy because it could result in 0 or 2+ increments. This code will get more complicated in the future (for #6109), so I pulled out the value filling logic to a separate method to isolate it from the increment logic. Also moved the value calculation logic down to where the values are filled since that's where it's used, and gave it a new name instead of overwriting the parameter.

Also update application_navigation_address_visibility_condition.test.ts to use value instead of values for its single value.