Fix JSON export with PAI
When getting the ReadOnlyApplicantProgramService inside the JsonExporterService, we are passing it an ApplicantData object without the ApplicantModel inside it. We're doing this intentionally so that any questions not in the application's program version are exported as "unanswered". This caused the lookup of data for a PAI name question (and probably any other PAI question) to fail since it relies on that ApplicantModel object.

This modifies the getters in ApplicantData to handle when the ApplicantModel is null.

In the future, we should probably figure out how to untangle this a bit to standardize how the ApplicantData object is created, but this fixes the immediate issue.