Use setting for APPLICANT_PORTAL_NAME instead of binding
Before we had the Settings system, we were binding APPLICANT_PORTAL_NAME directly from the config file into ApplicantAuthProviderName. Now that we have the settings system and need to pass in a Request object, this changes all the call sites to use the Settings getter instead (some had already made the change) and removes the now-unused binding.