Send email to guest applicant when PAI email question is answered
Previously, we were creating the ApplicantPersonalInfo object before we had actually submitted the application and saved the PAI data to the appropriate columns in the applicant table. This meant that guest applicants who answered a PAI email question never got the initial notification about applying to the program, including a default status message if it was set. This changes the logic to not create the ApplicantPersonalInfo object until after the application has been submitted.