Include North Star HTML files in prod image
When running in dev mode, since we volume mount the entire source path into the image, we have no problem referencing in the Thymeleaf HTML files. When we build the prod image, we are only copying in the files created by 'sbt dist'. Because we need to reference the HTML files directly in the code, we need to copy these files into the image. This maintains the same pathing present in the source so that we can reference the files in the same way, no matter if we're in dev or staging/prod mode.