Run the prod image locally with bin/run-dev --prod
Because the prod image is built prettty differently from the dev image,
we sometimes run into bugs that are only present on the prod image and
we don't see them when running the civiform-dev container with the full
source volume mounted into the container.

This provides a mechanism for using the civiform prod image with the dev
environment. It won't recompile the code when things change like the dev
image does, but it will allow us to poke at the prod image. It will run
in staging mode so that we still have access to dev tools.