Apply java options when forking is disabled
In order to attach a debugger to tests, we have to disable forking. When we disable forking, options defined in javaOptions are not passed through to the JVM, since we aren't forking a new JVM process. This can cause some things in tests to fail, such as using the testServerPort() helper. This change applies any Java options to the current JVM when forking is disabled. This only applies to system properties (-D) because we can't apply flags like Xmx or Xms, since the JVM is already running.